+++ CHANGELOG.md
@@ -58,3 +58,4 @@ All notable changes to the "my-racing-planner" project will be documented in thi
 - Added Hash Router using Wouter
 - Added Google analytics for page visits and outbound clicks
 - Added Privacy Policy page
+- Added instructions for PWA on Android and iOS